requirements.txt
asgiref==3.5.2
cachetools==5.2.0
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
Deprecated==1.2.13
Django==4.0.4
django-api-forms==0.7.1
django-cors-headers==3.12.0
django-crispy-forms==1.14.0
django-oauth-toolkit==2.1.0
djangorestframework==3.13.1
google-api-core==2.8.2
google-api-python-client==2.52.0
google-auth==2.9.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.3
httplib2==0.20.4
idna==3.3
jwcrypto==1.3.1
msgpack==1.0.0
oauthlib==3.2.0
protobuf==4.21.2
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2019.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.3.1
uritemplate==4.1.1
urllib3==1.26.9
utils==1.0.1
wrapt==1.14.1
